requirements.txt
streamlit
PyMuPDF
Pillow
transformers
torch
